[hiera-data] Add test user
I no user is specied in any hiera file, some of the modules I have
written myself. Like the emacs one will fail. The reason is that
it checks the accounts parameters to find where to put the emacs
configuration. You can remove the test user and replace it with
your own user and everything will work just fine.